1.添加了斑马线次数选择功能 2.优化了电机保护的阈值

+++ D_car_2025/D_car_2025/code/buzzer.c
@@ -8,7 +8,7 @@ void buzzer_init()
 }
 
 
-void buzzer_on(uint8 tim)
+void buzzer_on(uint16 tim)
 {
 	gpio_set_level(D7, GPIO_HIGH);
 	system_delay_ms(tim);
