+++ D_car_2025/D_car_2025/code/buzzer.h
@@ -3,7 +3,7 @@
 #include "zf_common_typedef.h"
 
 void buzzer_init();
-void buzzer_on(uint8 tim);
+void buzzer_on(uint16 tim);
